move autoplay tests to their own file
